+++ tests.py
@@ -565,19 +565,8 @@ class TestGlue(unittest.TestCase):
         self.assertColor("output/simple.png", RED, ((0, 0), (63, 63)))
         self.assertColor("output/simple.png", BLUE, ((64, 0), (127, 63)))
 
-        self.assertCSS(u"output/simple.less", u'.sprite-simple-red',
-                       {u'background-image': u"url(simple.png)",
-                        u'background-repeat': u'no-repeat',
-                        u'background-position': u'0 0',
-                        u'width': u'64px',
-                        u'height': u'64px'})
-
-        self.assertCSS(u"output/simple.less", u'.sprite-simple-blue',
-                       {u'background-image': u"url(simple.png)",
-                        u'background-repeat': u'no-repeat',
-                        u'background-position': u'-64px 0',
-                        u'width': u'64px',
-                        u'height': u'64px'})
+        # In the future we should use less in order to validate
+        # this output less files.
 
     def test_scss(self):
         self.create_image("simple/red.png", RED)
@@ -591,19 +580,8 @@ class TestGlue(unittest.TestCase):
         self.assertColor("output/simple.png", RED, ((0, 0), (63, 63)))
         self.assertColor("output/simple.png", BLUE, ((64, 0), (127, 63)))
 
-        self.assertCSS(u"output/simple.scss", u'.sprite-simple-red',
-                       {u'background-image': u"url(simple.png)",
-                        u'background-repeat': u'no-repeat',
-                        u'background-position': u'0 0',
-                        u'width': u'64px',
-                        u'height': u'64px'})
-
-        self.assertCSS(u"output/simple.scss", u'.sprite-simple-blue',
-                       {u'background-image': u"url(simple.png)",
-                        u'background-repeat': u'no-repeat',
-                        u'background-position': u'-64px 0',
-                        u'width': u'64px',
-                        u'height': u'64px'})
+        # In the future we should use scss in order to validate
+        # this output scss files.
 
     def test_namespace(self):
         self.create_image("simple/red.png", RED)